// frontend/src/types/supportRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SupportRequestType } from './supportRequestType';
import type { SupportRequestSeverity } from './supportRequestSeverity';

export interface SupportRequest {
  /** @minLength 1 */
  title: string;
  description?: string;
  type: SupportRequestType;
  severity: SupportRequestSeverity;
  attachmentIds?: number[];
}